+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 ### Added
 
 - Add player icon on map menu display from cc-multibakery
+- Add option to turn off main menu image replace
+- Add option to turn off inventory menu image replace
+- Add option to turn off map menu image replace
 
 ## [1.1.0] 2026-08-29